// src/generated/api.ts
export interface components {
    schemas: {
// ...
        Message: {
            /** @description Unique Id */
            _id: string;
            /** @description Unique value generated by client sending this message */
            nonce?: string | null;
            /** @description Id of the channel this message was sent in */
            channel: string;
            /** @description Id of the user or webhook that sent this message */
            author: string;
            /** @description The user that sent this message */
            user?: components["schemas"]["User"] | null;
            /** @description The member that sent this message */
            member?: components["schemas"]["Member"] | null;
            /** @description The webhook that sent this message */
            webhook?: components["schemas"]["MessageWebhook"] | null;
            /** @description Message content */
            content?: string | null;
            /** @description System message */
            system?: components["schemas"]["SystemMessage"] | null;
            /** @description Array of attachments */
            attachments?: components["schemas"]["File"][] | null;
            /** @description Time at which this message was last edited */
            edited?: components["schemas"]["ISO8601 Timestamp"] | null;
            /** @description Attached embeds to this message */
            embeds?: components["schemas"]["Embed"][] | null;
            /** @description Array of user ids mentioned in this message */
            mentions?: string[] | null;
            /** @description Array of role ids mentioned in this message */
            role_mentions?: string[] | null;
            /** @description Array of message ids this message is replying to */
            replies?: string[] | null;
            /** @description Hashmap of emoji IDs to array of user IDs */
            reactions?: {
                [key: string]: string[];
            };
            /** @description Information about how this message should be interacted with */
            interactions?: components["schemas"]["Interactions"];
            /** @description Name and / or avatar overrides for this message */
            masquerade?: components["schemas"]["Masquerade"] | null;
            /** @description Whether or not the message in pinned */
            pinned?: boolean | null;
            /**
             * Format: uint32
             * @description Bitfield of message flags
             *
             *     https://docs.rs/revolt-models/latest/revolt_models/v0/enum.MessageFlags.html
             */
            flags?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}